Refactor: HTTPレスポンスメッセージの各書式を関数ごとにファイルを作成した
変更理由：各書式ごとに拡張するため
