+++ srcs/http/response/create_crlf.cpp
@@ -1,0 +1,7 @@
+#include "character.hpp"
+#include <iostream>
+#include "http.hpp"
+
+void CreateCRLF(std::ostream &response_stream) {
+	response_stream << CR << LF;
+}
