+++ srcs/http/response/create_header_field.cpp
@@ -1,9 +1,10 @@
-#include <iostream>
 #include "character.hpp"
+#include <iostream>
+#include "http.hpp"
 
 void CreateHeaderField(std::ostream &response_stream) {
 	response_stream << "field-name"
-			  << ":" << SP << "field-value" << SP << CR << LF;
+					<< ":" << SP << "field-value" << SP << CR << LF;
 }
 
 void CreateHeaderFields(std::ostream &response_stream) {
